@@ -6,24 +6,23 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  transparent_image: ^2.0.0
-  intl: ^0.17.0
-  # http: ^0.13.1
-  json_annotation: ^3.1.1
   cupertino_icons: ^1.0.3
-  provider: ^4.3.3
+  intl: ^0.19.0
+  json_annotation: ^4.9.0
+  provider: ^6.1.2
+  transparent_image: ^2.0.0
   
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  pedantic: ^1.11.0
-  build_runner: ^1.10.9
-  json_serializable: ^3.5.1
+  build_runner: ^2.4.12
+  flutter_lints: ^4.0.0
+  json_serializable: ^6.8.0
 
 flutter:
   assets:
