@@ -1,14 +1,14 @@
 import 'package:carfax/data/account.dart';
-import 'package:carfax/pages/glovebox/glovebox_page.dart';
 import 'package:carfax/pages/dashboard/widgets/maintenance_item.dart';
 import 'package:carfax/pages/dashboard/widgets/recall_item.dart';
+import 'package:carfax/pages/glovebox/glovebox_page.dart';
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import 'package:provider/provider.dart';
 import 'package:transparent_image/transparent_image.dart';
 
 class DashboardPage extends StatefulWidget {
-  DashboardPage({Key key}) : super(key: key);
+  const DashboardPage({super.key});
 
   @override
   _DashboardPageState createState() => _DashboardPageState();
@@ -36,15 +36,15 @@ class _DashboardPageState extends State<DashboardPage> {
     return SizedBox(
       width: 80,
       child: Text(
-          '${NumberFormat.decimalPattern().format(context.watch<UserVehicle>().metric ? context.watch<UserVehicle>().kilometers : context.watch<UserVehicle>().miles)}',
+          NumberFormat.decimalPattern()
+              .format(context.watch<UserVehicle>().metric ? context.watch<UserVehicle>().kilometers : context.watch<UserVehicle>().miles),
           style: const TextStyle(fontSize: 16, color: Colors.blue)),
     );
   }
 
   Widget _odoEdit() {
-    _textController.text = context.watch<UserVehicle>().metric
-        ? context.read<UserVehicle>().kilometers.toString()
-        : context.read<UserVehicle>().miles.toString() ?? '';
+    _textController.text =
+        context.watch<UserVehicle>().metric ? context.read<UserVehicle>().kilometers.toString() : context.read<UserVehicle>().miles.toString();
     return SizedBox(
       width: 80,
       child: TextField(
@@ -56,8 +56,8 @@ class _DashboardPageState extends State<DashboardPage> {
         onSubmitted: (value) {
           if (_editingOdometer) {
             context.read<UserVehicle>().metric
-                ? context.read<UserVehicle>().kilometers = int.tryParse(value)
-                : context.read<UserVehicle>().miles = int.tryParse(value);
+                ? context.read<UserVehicle>().kilometers = int.parse(value)
+                : context.read<UserVehicle>().miles = int.parse(value);
           }
           setState(() {
             _editingOdometer = !_editingOdometer;
@@ -69,32 +69,25 @@ class _DashboardPageState extends State<DashboardPage> {
 
   Widget _buildOdometerEditor() {
     var metric = context.watch<UserVehicle>().metric;
-    return Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: <Widget>[
-          Text('Today\'s Odometer Reading ${metric ? '(KM)' : '(MI)'}',
-              style: const TextStyle(fontSize: 14)),
-          Row(mainAxisSize: MainAxisSize.min, children: [
-            _editingOdometer ? _odoEdit() : _odoText(),
-            IconButton(
-                icon: _editingOdometer
-                    ? const Icon(Icons.check)
-                    : const Icon(Icons.edit),
-                color: _editingOdometer ? Colors.green : Colors.blue,
-                onPressed: () {
-                  if (_editingOdometer) {
-                    metric
-                        ? context.read<UserVehicle>().kilometers =
-                            int.tryParse(_textController.value.text)
-                        : context.read<UserVehicle>().miles =
-                            int.tryParse(_textController.value.text);
-                  }
-                  setState(() {
-                    _editingOdometer = !_editingOdometer;
-                  });
-                }),
-          ])
-        ]);
+    return Column(crossAxisAlignment: CrossAxisAlignment.start, children: <Widget>[
+      Text('Today\'s Odometer Reading ${metric ? '(KM)' : '(MI)'}', style: const TextStyle(fontSize: 14)),
+      Row(mainAxisSize: MainAxisSize.min, children: [
+        _editingOdometer ? _odoEdit() : _odoText(),
+        IconButton(
+            icon: _editingOdometer ? const Icon(Icons.check) : const Icon(Icons.edit),
+            color: _editingOdometer ? Colors.green : Colors.blue,
+            onPressed: () {
+              if (_editingOdometer) {
+                metric
+                    ? context.read<UserVehicle>().kilometers = int.parse(_textController.value.text)
+                    : context.read<UserVehicle>().miles = int.parse(_textController.value.text);
+              }
+              setState(() {
+                _editingOdometer = !_editingOdometer;
+              });
+            }),
+      ])
+    ]);
   }
 
   Widget _buildOpenGlovebox() {
@@ -112,11 +105,8 @@ class _DashboardPageState extends State<DashboardPage> {
               MaterialPageRoute(
                 fullscreenDialog: true,
                 builder: (context) => MultiProvider(
-                  providers: [
-                    ChangeNotifierProvider.value(value: account),
-                    ChangeNotifierProvider.value(value: vehicle)
-                  ],
-                  child: GloveboxPage(),
+                  providers: [ChangeNotifierProvider.value(value: account), ChangeNotifierProvider.value(value: vehicle)],
+                  child: const GloveboxPage(),
                 ),
               ),
             );
@@ -129,36 +119,31 @@ class _DashboardPageState extends State<DashboardPage> {
 
   Widget _buildVehicleHeader() {
     return Card(
-      child: Row(
-          mainAxisAlignment: MainAxisAlignment.spaceAround,
-          mainAxisSize: MainAxisSize.max,
-          children: [
-            FadeInImage.memoryNetwork(
-              placeholder: kTransparentImage,
-              height: 100,
-              width: 100,
-              fit: BoxFit.contain,
-              image: context.read<UserVehicle>().imageURL,
-            ),
-            _buildOdometerEditor(),
-            _buildOpenGlovebox(),
-          ]),
+      child: Row(mainAxisAlignment: MainAxisAlignment.spaceAround, mainAxisSize: MainAxisSize.max, children: [
+        FadeInImage.memoryNetwork(
+          placeholder: kTransparentImage,
+          height: 100,
+          width: 100,
+          fit: BoxFit.contain,
+          image: context.read<UserVehicle>().imageURL,
+        ),
+        _buildOdometerEditor(),
+        _buildOpenGlovebox(),
+      ]),
     );
   }
 
   Future<void> _loadMaintenanceItems() async {
     final items = [
-      MaintenanceItem(const Icon(Icons.local_gas_station), 'Oil Change', 7, 8),
-      MaintenanceItem(
-          const Icon(Icons.text_snippet_rounded), 'Registration', 11, 24),
-      MaintenanceItem(
-          const Icon(Icons.miscellaneous_services_sharp), 'Tread Life', 2, 4),
-      RecallItem('No Open Recalls Reported')
+      const MaintenanceItem(Icon(Icons.local_gas_station), 'Oil Change', 7, 8),
+      const MaintenanceItem(Icon(Icons.text_snippet_rounded), 'Registration', 11, 24),
+      const MaintenanceItem(Icon(Icons.miscellaneous_services_sharp), 'Tread Life', 2, 4),
+      const RecallItem('No Open Recalls Reported')
     ];
     for (var item in items) {
-      await Future.delayed(Duration(milliseconds: 1));
+      await Future.delayed(const Duration(milliseconds: 1));
       _listItems.add(item);
-      _listKey.currentState.insertItem(_listItems.length - 1);
+      _listKey.currentState?.insertItem(_listItems.length - 1);
     }
   }
 
@@ -170,20 +155,17 @@ class _DashboardPageState extends State<DashboardPage> {
           padding: const EdgeInsets.only(top: 10),
           initialItemCount: _listItems.length,
           itemBuilder: (context, index, animation) {
-            return SizeTransition(
-                axis: Axis.vertical,
-                sizeFactor: animation,
-                child: _listItems[index]);
+            return SizeTransition(axis: Axis.vertical, sizeFactor: animation, child: _listItems[index]);
           }),
     );
   }
 
   Widget _buildHelpLink() {
-    return Padding(
-      padding: const EdgeInsets.all(10),
+    return const Padding(
+      padding: EdgeInsets.all(10),
       child: Row(mainAxisAlignment: MainAxisAlignment.center, children: [
-        const Text('Something Wrong?'),
-        const Text(' Get Help', style: TextStyle(color: Colors.blue)),
+        Text('Something Wrong?'),
+        Text(' Get Help', style: TextStyle(color: Colors.blue)),
       ]),
     );
   }
