@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
 class MaintenanceItem extends StatefulWidget {
-  MaintenanceItem(this.icon, this.title, this.durationLeft, this.durationTotal);
+  const MaintenanceItem(this.icon, this.title, this.durationLeft, this.durationTotal, {super.key});
 
   final Icon icon;
   final String title;
@@ -11,20 +11,17 @@ class MaintenanceItem extends StatefulWidget {
   _MaintenanceItemState createState() => _MaintenanceItemState();
 }
 
-class _MaintenanceItemState extends State<MaintenanceItem>
-    with TickerProviderStateMixin {
-  AnimationController _controller;
+class _MaintenanceItemState extends State<MaintenanceItem> with TickerProviderStateMixin {
+  late AnimationController _controller;
 
   @override
   void initState() {
+    super.initState();
     _controller = AnimationController(
       vsync: this,
       duration: const Duration(seconds: 2),
-    )..addListener(() {
-        setState(() {});
-      });
+    );
     _controller.animateTo(1 - (widget.durationLeft / widget.durationTotal));
-    super.initState();
   }
 
   @override
@@ -44,11 +41,8 @@ class _MaintenanceItemState extends State<MaintenanceItem>
           Expanded(flex: 1, child: widget.icon),
           Expanded(
             flex: 5,
-            child:
-                Column(crossAxisAlignment: CrossAxisAlignment.start, children: [
-              Text(widget.title,
-                  style: const TextStyle(
-                      fontSize: 16, fontWeight: FontWeight.bold)),
+            child: Column(crossAxisAlignment: CrossAxisAlignment.start, children: [
+              Text(widget.title, style: const TextStyle(fontSize: 16, fontWeight: FontWeight.bold)),
               Padding(
                 padding: const EdgeInsets.only(top: 5),
                 child: LinearProgressIndicator(value: _controller.value),
@@ -57,12 +51,11 @@ class _MaintenanceItemState extends State<MaintenanceItem>
           ),
           Expanded(
             flex: 3,
-            child: Column(mainAxisSize: MainAxisSize.min, children: <Widget>[
-              Text(widget.durationLeft.toString()),
-              Text('$durationText Left', style: const TextStyle(fontSize: 12))
-            ]),
+            child: Column(
+                mainAxisSize: MainAxisSize.min,
+                children: <Widget>[Text(widget.durationLeft.toString()), Text('$durationText Left', style: const TextStyle(fontSize: 12))]),
           ),
-          Expanded(flex: 0, child: const Icon(Icons.arrow_right))
+          const Expanded(flex: 0, child: Icon(Icons.arrow_right))
         ]),
       ),
       const Divider()
