@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
 class DistanceDrivenDropdown extends StatefulWidget {
-  DistanceDrivenDropdown({Key key}) : super(key: key);
+  const DistanceDrivenDropdown({super.key});
 
   @override
   _DistanceDrivenDropdownState createState() => _DistanceDrivenDropdownState();
@@ -14,16 +14,15 @@ class _DistanceDrivenDropdownState extends State<DistanceDrivenDropdown> {
     return DropdownButton<String>(
       isExpanded: true,
       value: _dropdownValue,
-      style: TextStyle(fontSize: 12),
-      onChanged: (String newValue) {
-        setState(() => _dropdownValue = newValue);
+      style: const TextStyle(fontSize: 12),
+      onChanged: (newValue) {
+        setState(() => _dropdownValue = newValue!);
       },
       items: <String>[
         'Car Care Calculated',
         'My Estimate',
       ].map<DropdownMenuItem<String>>((String value) {
-        return DropdownMenuItem<String>(
-            value: value, child: Text(value, style: TextStyle(fontSize: 12)));
+        return DropdownMenuItem<String>(value: value, child: Text(value, style: const TextStyle(fontSize: 12)));
       }).toList(),
     );
   }
