@@ -1,13 +1,13 @@
 import 'package:carfax/data/account.dart';
+import 'package:carfax/data/bottom_nav_bar.dart';
 import 'package:carfax/pages/dashboard/dashboard_page.dart';
 import 'package:carfax/pages/history/history_page.dart';
-import 'package:carfax/data/bottom_nav_bar.dart';
 import 'package:carfax/utilities/get_utility.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
 class VehicleDetailPage extends StatefulWidget {
-  VehicleDetailPage({Key key, @required this.vin}) : super(key: key);
+  const VehicleDetailPage({super.key, required this.vin});
 
   final String vin;
 
@@ -23,13 +23,10 @@ class _VehicleDetailPageState extends State<VehicleDetailPage> {
   @override
   void initState() {
     super.initState();
-    GetUtility.getVehicleDetails(widget.vin).then((value) {
-      if (value != null) {
-        context.read<UserVehicle>().bodyTypeDescription =
-            value.bodyTypeDescription;
-        context.read<UserVehicle>().events = value.events;
-        setState(() => _loading = false);
-      }
+    GetUtility().getVehicleDetails(widget.vin).then((value) {
+      context.read<UserVehicle>().bodyTypeDescription = value!.bodyTypeDescription;
+      context.read<UserVehicle>().events = value.events;
+      setState(() => _loading = false);
     });
   }
 
@@ -43,8 +40,7 @@ class _VehicleDetailPageState extends State<VehicleDetailPage> {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text(
-            '${context.read<UserVehicle>().year} ${context.read<UserVehicle>().make} ${context.read<UserVehicle>().model}'),
+        title: Text('${context.read<UserVehicle>().year} ${context.read<UserVehicle>().make} ${context.read<UserVehicle>().model}'),
         centerTitle: true,
       ),
       body: _loading
@@ -54,13 +50,10 @@ class _VehicleDetailPageState extends State<VehicleDetailPage> {
               onPageChanged: (int idx) {
                 setState(() => _selectedIndex = idx);
               },
-              children: <Widget>[DashboardPage(), HistoryPage()]),
+              children: const <Widget>[DashboardPage(), HistoryPage()]),
       bottomNavigationBar: BottomNavigationBar(
           items: bottomNavItems.map((BottomNav bottomNavMenuItem) {
-            return BottomNavigationBarItem(
-                icon: Icon(bottomNavMenuItem.icon),
-                label: bottomNavMenuItem.title,
-                backgroundColor: Theme.of(context).primaryColor);
+            return BottomNavigationBarItem(icon: Icon(bottomNavMenuItem.icon), label: bottomNavMenuItem.title, backgroundColor: Theme.of(context).primaryColor);
           }).toList(),
           currentIndex: _selectedIndex,
           onTap: (int idx) {
