@@ -1,14 +1,15 @@
 import 'package:carfax/data/account.dart';
 import 'package:carfax/pages/home/home_page.dart';
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
 void main() {
-  runApp(MainPage());
+  runApp(const MainPage());
 }
 
 class MainPage extends StatelessWidget {
+  const MainPage({super.key});
+
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
@@ -16,7 +17,7 @@ class MainPage extends StatelessWidget {
       theme: ThemeData(primarySwatch: Colors.blue, brightness: Brightness.dark),
       home: ChangeNotifierProvider(
         create: (_) => UserAccount(),
-        child: HomePage(),
+        child: const HomePage(),
       ),
     );
   }
