@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 
 class RecallItem extends StatelessWidget {
-  RecallItem(this.title);
+  const RecallItem(this.title, {super.key});
 
   final String title;
 
@@ -12,29 +12,24 @@ class RecallItem extends StatelessWidget {
       children: [
         Padding(
           padding: const EdgeInsets.all(10),
-          child:
-              Row(mainAxisAlignment: MainAxisAlignment.spaceAround, children: [
-            Expanded(flex: 1, child: Icon(Icons.check)),
+          child: Row(mainAxisAlignment: MainAxisAlignment.spaceAround, children: [
+            const Expanded(flex: 1, child: Icon(Icons.check)),
             Expanded(
               flex: 8,
-              child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Text(
-                      title,
-                      style:
-                          TextStyle(fontSize: 16, fontWeight: FontWeight.bold),
-                    ),
-                    Padding(
-                      padding: const EdgeInsets.only(top: 5),
-                      child: Text(
-                          'As of ${DateFormat('dd/MM/yyyy', 'en_US').format(DateTime.now())}'),
-                    )
-                  ]),
+              child: Column(crossAxisAlignment: CrossAxisAlignment.start, children: [
+                Text(
+                  title,
+                  style: const TextStyle(fontSize: 16, fontWeight: FontWeight.bold),
+                ),
+                Padding(
+                  padding: const EdgeInsets.only(top: 5),
+                  child: Text('As of ${DateFormat('dd/MM/yyyy', 'en_US').format(DateTime.now())}'),
+                )
+              ]),
             )
           ]),
         ),
-        Divider()
+        const Divider()
       ],
     );
   }
