@@ -1,20 +1,21 @@
 import 'dart:convert';
+
 import 'package:carfax/data/account.dart';
 import 'package:carfax/json/account_json.dart';
 import 'package:carfax/json/vehicle_details_json.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/foundation.dart';
-import 'package:provider/provider.dart';
 import 'package:flutter/services.dart' show rootBundle;
+import 'package:provider/provider.dart';
 
 class GetUtility {
-  static Future<void> getAccount(BuildContext context) async {
+  Future<void> getAccount(BuildContext context) async {
     try {
       var data = await rootBundle.loadString('assets/data/account.json');
       var account = await compute(parseAccountJson, data);
       context.read<UserAccount>().email = account.email;
       context.read<UserAccount>().isCanadian = account.isCanadian;
-      account.vehicles.forEach((element) {
+      for (var element in account.vehicles) {
         var userVehicle = UserVehicle();
         userVehicle.make = element.make;
         userVehicle.model = element.model;
@@ -28,46 +29,44 @@ class GetUtility {
         userVehicle.kilometers = element.lastOdoKm;
         userVehicle.miles = element.lastOdoMileage;
         context.read<UserAccount>().vehicles.add(userVehicle);
-      });
+      }
     } catch (e, s) {
       print('$e $s');
     }
   }
 
-  static AccountJson parseAccountJson(String data) {
-    return AccountJson.fromJson(json.decode(data));
+  AccountData parseAccountJson(String data) {
+    return AccountData.fromJson(json.decode(data));
   }
 
-  static Future<VehicleDetailsJson> getVehicleDetails(String vin) async {
+  Future<VehicleDetails?> getVehicleDetails(String vin) async {
     try {
-      var data =
-          await rootBundle.loadString('assets/data/vehicle_details.json');
+      var data = await rootBundle.loadString('assets/data/vehicle_details.json');
       return compute(parseVehicleDetailsJson, data);
     } catch (e, s) {
       print('$e $s');
-      return null;
+      throw Exception();
     }
   }
 
-  static VehicleDetailsJson parseVehicleDetailsJson(String data) {
-    return VehicleDetailsJson.fromJson(json.decode(data));
+  VehicleDetails parseVehicleDetailsJson(String data) {
+    return VehicleDetails.fromJson(json.decode(data));
   }
 
-  static Future<List<ServiceRecord>> getShopRecords() async {
+  Future<List<ServiceRecord>> getShopRecords() async {
     try {
-      var data =
-          await rootBundle.loadString('assets/data/vehicle_details.json');
+      var data = await rootBundle.loadString('assets/data/vehicle_details.json');
       return compute(parseShopRecordsJson, data);
     } catch (e, s) {
       print('$e $s');
-      return null;
+      return [];
     }
   }
 
-  static List<ServiceRecord> parseShopRecordsJson(String data) {
-    var vehicleDetails = VehicleDetailsJson.fromJson(json.decode(data));
+  List<ServiceRecord> parseShopRecordsJson(String data) {
+    var vehicleDetails = VehicleDetails.fromJson(json.decode(data));
     var serviceRecords = <ServiceRecord>[];
-    vehicleDetails.displayRecords.forEach((element) {
+    for (var element in vehicleDetails.displayRecords) {
       var serviceRecord = ServiceRecord();
       serviceRecord.shopName = element.source.first.text;
       serviceRecord.kmAtService = element.odometerKm;
@@ -75,7 +74,7 @@ class GetUtility {
       serviceRecord.date = element.displayDate;
       serviceRecord.servicePerformed = element.details.last.text;
       serviceRecords.add(serviceRecord);
-    });
+    }
     return serviceRecords;
   }
 }
